requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
attrs==23.2.0
bitarray==2.9.2
certifi==2024.2.2
charset-normalizer==3.3.2
crontab==1.0.1
cytoolz==0.12.3
eth-account==0.11.0
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==4.0.0
eth_abi==5.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multidict==6.0.5
parsimonious==0.9.0
protobuf==5.26.0
psycopg2==2.9.9
pycryptodome==3.20.0
python-dotenv==1.0.1
python-telegram-bot==21.0.1
pyunormalize==15.1.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.18.0
sniffio==1.3.1
SQLAlchemy==2.0.29
toolz==0.12.1
typing_extensions==4.10.0
urllib3==2.2.1
web3==6.15.1
websockets==12.0
yarl==1.9.4
